Level 0, plus more particles and SFX
Added Level 0 for an even easier first level to the game. Changed the progression to move through levels and keep score. Added more particles to everything. SFX for cauldron slash.

+++ Witch-Soup/Assets/Scripts/VeggieLaunched.cs
@@ -24,7 +24,7 @@ public class VeggieLaunched : MonoBehaviour
     }
 
     public void getClicked() {
-        UIManager.score += 10 + Mathf.Floor(transform.localPosition.y);
+        UIManager.score += 10 + (int)Mathf.Floor(transform.localPosition.y);
         Destroy(gameObject);
     }
 }
